requirements.txt
pyTelegramBotAPI==4.14.0
Flask==2.3.3
APScheduler==3.10.1
python-dotenv==1.0.0
gunicorn==21.2.0